+++ cli/src/commands/build.rs
@@ -4,21 +4,32 @@ use clap::ArgMatches;
 use derive_more::Display;
 use derive_more::Error;
 use log::*;
+use serde::Serialize;
 use smaug::dragonruby;
 use smaug::util::dir::copy_directory;
 use std::env;
 use std::path::Path;
+use std::path::PathBuf;
 use std::process;
 
 #[derive(Debug)]
 pub struct Build;
 
-#[derive(Debug, Display, Error)]
+#[derive(Debug, Serialize, Display)]
+#[display(fmt = "Successfully built {}", "project_name")]
+pub struct BuildResult {
+    project_name: String,
+}
+#[derive(Debug, Display, Error, Serialize)]
 pub enum Error {
     #[display(
         fmt = "Could not find the configured version of DragonRuby. Install it with `smaug dragonruby install`"
     )]
     ConfiguredDragonRubyNotFound,
+    #[display(fmt = "Couldn't load Smaug configuration.")]
+    ConfigError { path: PathBuf },
+    #[display(fmt = "Could not find file at {}", "path.display()")]
+    FileNotFound { path: PathBuf },
 }
 
 impl Command for Build {
@@ -35,12 +46,21 @@ impl Command for Build {
             .value_of("path")
             .unwrap_or_else(|| current_directory.to_str().unwrap());
         debug!("Directory: {}", directory);
-        let path = Path::new(directory);
-        let path = std::fs::canonicalize(&path).expect("Could not find path");
+        let path = match std::fs::canonicalize(directory) {
+            Ok(dir) => dir,
+            Err(..) => {
+                return Err(Box::new(Error::FileNotFound {
+                    path: Path::new(directory).to_path_buf(),
+                }))
+            }
+        };
 
         let config_path = path.join("Smaug.toml");
 
-        let config = smaug::config::load(&config_path)?;
+        let config = match smaug::config::load(&config_path) {
+            Ok(conf) => conf,
+            Err(..) => return Err(Box::new(Error::ConfigError { path: config_path })),
+        };
         debug!("Smaug config: {:?}", config);
 
         trace!("Writing game metadata.");
@@ -75,11 +95,21 @@ impl Command for Build {
                     path.to_str().unwrap(),
                     dragonruby_options.join(" "),
                 );
+
+                let quiet = matches.is_present("json") || matches.is_present("quiet");
+
+                let stdout = if quiet {
+                    process::Stdio::null()
+                } else {
+                    process::Stdio::inherit()
+                };
+
                 process::Command::new(bin)
                     .current_dir(bin_dir.to_str().unwrap())
                     .arg("--only-package")
                     .args(dragonruby_options)
                     .arg(path.file_name().unwrap())
+                    .stdout(stdout)
                     .spawn()
                     .unwrap()
                     .wait()
@@ -107,10 +137,9 @@ impl Command for Build {
                         .expect("couldn't copy exceptions");
                 }
 
-                Ok(Box::new(format!(
-                    "Successfully built {} to the builds dir.",
-                    config.project.unwrap().name
-                )))
+                Ok(Box::new(BuildResult {
+                    project_name: config.project.unwrap().name,
+                }))
             }
         }
     }