+++ cli/src/commands/dragonruby/uninstall.rs
@@ -2,56 +2,46 @@ use crate::command::Command;
 use crate::command::CommandResult;
 use ::smaug::dragonruby;
 use clap::ArgMatches;
+use derive_more::Display;
+use derive_more::Error;
 use log::*;
+use serde::Serialize;
 use smaug::smaug;
-
-use super::list::List;
 
 #[derive(Debug)]
 pub struct Uninstall;
 
+#[derive(Debug, Display, Serialize)]
+#[display(fmt = "Uninstalled DragonRuby {}", "version")]
+pub struct UninstallResult {
+    version: String,
+}
+
+#[derive(Debug, Display, Error, Serialize)]
+enum Error {
+    #[display(fmt = "Could not find DragonRuby {}", "version")]
+    DragonRubyNotFound { version: String },
+}
+
 impl Command for Uninstall {
     fn run(&self, matches: &ArgMatches) -> CommandResult {
         trace!("Uninstall Command");
-        let version = matches.value_of("VERSION");
+        let version = matches.value_of("VERSION").expect("No version specified");
 
-        match version {
-            None => {
-                let list = List.run(matches);
+        let location = smaug::data_dir().join("dragonruby").join(version);
+        let dragonruby = dragonruby::new(&location);
 
-                match list {
-                    Ok(message) => Ok(Box::new(format!(
-                        "smaug dragonruby list <VERSION>\n\nHere are the installed versions:\n\n{}",
-                        message
-                    ))),
-                    Err(..) => list,
-                }
+        match dragonruby {
+            Ok(dr) => {
+                rm_rf::ensure_removed(dr.path).expect("DragonRuby folder not removed.");
+
+                Ok(Box::new(UninstallResult {
+                    version: version.to_string(),
+                }))
             }
-            Some(version) => {
-                let location = smaug::data_dir().join("dragonruby").join(version);
-                let dragonruby = dragonruby::new(&location);
-
-                match dragonruby {
-                    Ok(dr) => {
-                        rm_rf::ensure_removed(dr.path).expect("DragonRuby folder not removed.");
-                        let message = format!("Uninstalled {}", dr.version);
-
-                        Ok(Box::new(message))
-                    }
-                    Err(..) => {
-                        let list = List.run(matches);
-
-                        match list {
-                            Ok(message) => Ok(Box::new(format!(
-                        "DragonRuby version \"{}\" not found. Here are the installed versions:\n\n{}",
-                        version,
-                        message
-                    ))),
-                            Err(..) => list,
-                        }
-                    }
-                }
-            }
+            Err(..) => Err(Box::new(Error::DragonRubyNotFound {
+                version: version.to_string(),
+            })),
         }
     }
 }