+++ cli/src/commands/package.rs
@@ -11,14 +11,11 @@ pub struct Package;
 impl Command for Package {
     fn run(&self, matches: &ArgMatches) -> CommandResult {
         trace!("Package Command");
-
-        let command: Box<dyn Command> = match matches.subcommand_name() {
-            Some("init") => Box::new(init::Init),
-            _ => unreachable!(),
-        };
-
         let subcommand_matches = matches.subcommand_matches(matches.subcommand_name().unwrap());
 
-        command.run(&subcommand_matches.unwrap())
+        match matches.subcommand_name() {
+            Some("init") => init::Init.run(&subcommand_matches.unwrap()),
+            _ => unreachable!(),
+        }
     }
 }