+++ smaug/src/dependency.rs
@@ -1,4 +1,6 @@
-#[derive(Clone, Debug)]
+use serde::Serialize;
+
+#[derive(Clone, Debug, Serialize)]
 pub struct Dependency {
     pub name: String,
     pub version: String,