+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+# Version 0.3.0
+
+* Add `--json` argument to all commands
+
 # Version 0.2.3
 
-* Fixed dragonruby install from itch.io package
-* Fixed issue with add with empty dependencies in config
-* Fixed `smaug.rb` generation on Windows
+* Fix dragonruby install from itch.io package
+* Fix issue with add with empty dependencies in config
+* Fix `smaug.rb` generation on Windows
 * Add `compile_ruby` flag to enable Bytecode compilation
 * Automatically run `smaug install` on `smaug add`